testProject/requirements.txt
Django
djangorestframework
Djongo
pymongo==3.12.3
pyjwt